struct TERRAIN_INFO : use "uint32_t" instead of "long" for rows columns members. sizeoff(long) are platform dependent, this struct is read from Terrain file (.DEM) so for interoperability, we can't use platform dependent sized type !

+++ Common/Header/RasterTerrain.h
@@ -11,8 +11,8 @@ typedef struct _TERRAIN_INFO
   double Top;
   double Bottom;
   double StepSize;
-  long Rows;
-  long Columns;
+  uint32_t Rows;
+  uint32_t Columns;
 } TERRAIN_INFO;
